requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.11.0

# Reinforcement Learning
gym>=0.21.0
gymnasium>=0.28.0
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# AI Agents & LLM Integration
openai>=1.3.0
pyautogen>=0.2.0
autogenstudio>=0.1.0
autogen-ext[openai]>=0.1.0
autogen-core>=0.1.0
autogen-agentchat>=0.1.0

# Data handling
yfinance>=0.1.70
alpha-vantage>=2.3.1
pandas-ta>=0.3.14b
ib_insync>=0.9.70
ta>=0.10.2

# Feature Store dependencies
duckdb>=0.9.0
zstandard>=0.21.0
pyarrow>=10.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter for high-concurrency manifest operations

# Survivorship Bias Analysis
sqlite3  # Built into Python
openpyxl>=3.0.0  # For Excel file support

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.6.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
requests>=2.27.0
redis>=5.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.20

# Monitoring & Metrics
prometheus-client>=0.17.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
ruff>=0.1.0
mypy>=1.0.0
isort>=5.10.0
yamllint>=1.32.0  # Keep config files clean

# Additional ML/RL dependencies
tensorboard>=2.8.0
wandb>=0.15.0
optuna>=3.0.0